+++ features/merchant/product_detail_common/src/main/java/com/tokopedia/product/detail/common/data/model/variant/VariantChild.kt
@@ -15,10 +15,9 @@ data class VariantChild(
         @Expose
         val productId: String = "", //ex: 15212348
 
-
         @SerializedName("price")
         @Expose
-        val price: java.lang.Double = (0.0 as java.lang.Double), //ex: 100000000
+        val price: Double = 0.0, //ex: 100000000
 
         @SerializedName("priceFmt")
         @Expose
@@ -74,9 +73,9 @@ data class VariantChild(
     val finalMainPrice: Double
         get() {
             return if (campaign?.isActive == true) {
-                campaign.originalPrice ?: (0.0 as Double)
+                campaign.originalPrice ?: 0.0
             } else {
-                price.toDouble()
+                price
             }
         }
 
@@ -85,8 +84,7 @@ data class VariantChild(
             return if (campaign?.isActive == true) {
                 campaign.discountedPrice ?: 0.0
             } else {
-                price.toDouble()
-
+                price
             }
         }
