+++ features/merchant/product_manage_common/src/main/java/com/tokopedia/product/manage/common/feature/variant/data/mapper/ProductManageVariantMapper.kt
@@ -34,7 +34,7 @@ object ProductManageVariantMapper {
                 it.combination,
                 it.isPrimary,
                 it.isCampaign,
-                it.price.toDouble(),
+                it.price,
                 it.sku,
                 it.stock,
                 it.pictures,
